backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
deepl==1.19.1
distro==1.9.0
faiss-cpu==1.9.0
fastapi==0.115.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.20.1
kiwipiepy_model==0.20.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-teddynote==0.3.21
langchain-text-splitters==0.3.2
langgraph==0.2.44
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.139
lz4==4.3.3
marshmallow==3.23.1
mmh3==4.1.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.53.0
orjson==3.10.11
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
propcache==0.2.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyMuPDF==1.24.13
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.3.0
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.6
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wget==3.2
wheel==0.44.0
yarl==1.17.1
